refactor configurator (ba72586e9)

+++ ante/testutils/testutil.go
@@ -110,7 +110,7 @@ func (suite *AnteTestSuite) SetupTest() {
 	configurator.ResetTestChainConfig()
 	err := configurator.
 		WithChainConfig(chainConfig).
-		WithEVMCoinInfo(denom, decimals).
+		WithEVMCoinInfo(denom, uint8(decimals)).
 		Configure()
 	suite.Require().NoError(err)
