+++ example_chain/config.go
@@ -34,7 +34,7 @@ func InitializeAppConfiguration(chainID string) error {
 	err = evmtypes.NewEVMConfigurator().
 		WithChainConfig(ethCfg).
 		// NOTE: we're using the 18 decimals default for the example chain
-		WithEVMCoinInfo(baseDenom, evmtypes.EighteenDecimals).
+		WithEVMCoinInfo(baseDenom, uint8(evmtypes.EighteenDecimals)).
 		Configure()
 	if err != nil {
 		return err
