+++ example_chain/config_testing.go
@@ -37,7 +37,7 @@ func InitializeAppConfiguration(chainID string) error {
 	err = configurator.
 		WithExtendedEips(evmosActivators).
 		WithChainConfig(ethCfg).
-		WithEVMCoinInfo(baseDenom, evmtypes.EighteenDecimals).
+		WithEVMCoinInfo(baseDenom, uint8(evmtypes.EighteenDecimals)).
 		Configure()
 	if err != nil {
 		return err
